+++ sdk/src/test/java/com/iovation/launchkey/sdk/client/BasicServiceClientGetAuthorizationResponseTest.java
@@ -1,10 +1,7 @@
 package com.iovation.launchkey.sdk.client;
 
+import com.iovation.launchkey.sdk.domain.service.AuthMethod.Type;
 import com.iovation.launchkey.sdk.domain.service.AuthorizationResponse;
-import com.iovation.launchkey.sdk.domain.webhook.AuthorizationResponseWebhookPackage;
-import com.iovation.launchkey.sdk.domain.webhook.ServiceUserSessionEndWebhookPackage;
-import com.iovation.launchkey.sdk.domain.webhook.WebhookPackage;
-import com.iovation.launchkey.sdk.error.InvalidRequestException;
 import com.iovation.launchkey.sdk.transport.Transport;
 import com.iovation.launchkey.sdk.transport.domain.*;
 import junit.framework.TestCase;
@@ -13,12 +10,14 @@ import org.junit.Test;
 import org.junit.experimental.runners.Enclosed;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
-import org.mockito.ArgumentMatchers;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 
 import java.util.*;
 
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.collection.IsEmptyCollection.empty;
 import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
@@ -54,7 +53,16 @@ public class BasicServiceClientGetAuthorizationResponseTest extends TestCase {
                     new String[]{"service", "pins"},
                     "type",
                     "reason",
-                    "denial-reason"
+                    "denial-reason",
+                    new AuthPolicy(
+                            1, true, true, true, null
+                    ),
+                    new AuthMethod[]{
+                            new AuthMethod("meth-true", true, true, true, true, true, true, true),
+                            new AuthMethod("meth-false", false, false, false, false, false, false, false),
+                            new AuthMethod("meth-null", null, null, null, null, null, null, null)
+
+                    }
             );
             when(transport.serviceV3AuthsGet(any(UUID.class), any(EntityIdentifier.class))).thenReturn(authResponse);
         }
@@ -120,6 +128,14 @@ public class BasicServiceClientGetAuthorizationResponseTest extends TestCase {
             assertEquals("denial-reason", response.getDenialReason());
         }
 
+        @Test
+        public void testNullTransportResponseReturnsNull() throws Exception {
+            when(transport.serviceV3AuthsGet(any(UUID.class), any(EntityIdentifier.class))).thenReturn(null);
+            AuthorizationResponse response = client.getAuthorizationResponse(authRequestId.toString());
+            assertNull(response);
+        }
+
+
     }
 
     @RunWith(Parameterized.class)
@@ -157,8 +173,16 @@ public class BasicServiceClientGetAuthorizationResponseTest extends TestCase {
                     new String[]{"service", "pins"},
                     this.input,
                     "reason",
-                    "denial-reason"
-            ));
+                    "denial-reason",
+                    new AuthPolicy(
+                            1, true, true, true, null
+                    ),
+                    new AuthMethod[]{
+                            new AuthMethod("meth-true", true, true, true, true, true, true, true),
+                            new AuthMethod("meth-false", false, false, false, false, false, false, false),
+                            new AuthMethod("meth-null", null, null, null, null, null, null, null)
+
+                    }            ));
             ServiceClient client = new BasicServiceClient(UUID.randomUUID(), transport);
             AuthorizationResponse response = client.getAuthorizationResponse(UUID.randomUUID().toString());
             assertEquals(this.expectedOutput, response.getType());
@@ -180,6 +204,8 @@ public class BasicServiceClientGetAuthorizationResponseTest extends TestCase {
                     { "PERMISSION", AuthorizationResponse.Reason.PERMISSION },
                     { "AUTHENTICATION", AuthorizationResponse.Reason.AUTHENTICATION },
                     { "CONFIGURATION", AuthorizationResponse.Reason.CONFIGURATION },
+                    { "BUSY_LOCAL", AuthorizationResponse.Reason.BUSY_LOCAL },
+                    { "SENSOR", AuthorizationResponse.Reason.SENSOR },
                     { "UNKNOWN", AuthorizationResponse.Reason.OTHER },
             });
         }
@@ -204,7 +230,16 @@ public class BasicServiceClientGetAuthorizationResponseTest extends TestCase {
                     new String[]{"service", "pins"},
                     "type",
                     this.input,
-                    "denial-reason"
+                    "denial-reason",
+                    new AuthPolicy(
+                            1, true, true, true, null
+                    ),
+                    new AuthMethod[]{
+                            new AuthMethod("meth-true", true, true, true, true, true, true, true),
+                            new AuthMethod("meth-false", false, false, false, false, false, false, false),
+                            new AuthMethod("meth-null", null, null, null, null, null, null, null)
+
+                    }
             ));
             ServiceClient client = new BasicServiceClient(UUID.randomUUID(), transport);
             AuthorizationResponse response = client.getAuthorizationResponse(UUID.randomUUID().toString());
@@ -251,7 +286,16 @@ public class BasicServiceClientGetAuthorizationResponseTest extends TestCase {
                     new String[]{"service", "pins"},
                     "type",
                     this.input,
-                    "denial-reason"
+                    "denial-reason",
+                    new AuthPolicy(
+                            1, true, true, true, null
+                    ),
+                    new AuthMethod[]{
+                            new AuthMethod("meth-true", true, true, true, true, true, true, true),
+                            new AuthMethod("meth-false", false, false, false, false, false, false, false),
+                            new AuthMethod("meth-null", null, null, null, null, null, null, null)
+
+                    }
             ));
             ServiceClient client = new BasicServiceClient(UUID.randomUUID(), transport);
             AuthorizationResponse response = client.getAuthorizationResponse(UUID.randomUUID().toString());
@@ -259,4 +303,356 @@ public class BasicServiceClientGetAuthorizationResponseTest extends TestCase {
         }
     }
 
+    @RunWith(MockitoJUnitRunner.class)
+    public static class PolicyTests {
+
+        @Mock
+        private Transport transport;
+
+        @Mock
+        private ServiceV3AuthsGetResponse transportResponse;
+
+        private BasicServiceClient client;
+
+        private UUID serviceId = UUID.randomUUID();
+
+        private UUID authRequestId = UUID.randomUUID();
+
+        @Before
+        public void setUp() throws Exception {
+            when(transport.serviceV3AuthsGet(any(UUID.class), any(EntityIdentifier.class))).thenReturn(transportResponse);
+            when(transportResponse.getAuthorizationRequestId()).thenReturn(authRequestId);
+            when(transportResponse.getServicePins()).thenReturn(new String[]{});
+            client = new BasicServiceClient(serviceId, transport);
+        }
+
+        @Test
+        public void noPolicyReturnsNull() throws Exception {
+            when(transportResponse.getAuthPolicy()).thenReturn(null);
+            AuthorizationResponse response = client.getAuthorizationResponse(authRequestId.toString());
+            assertNull(response.getPolicy());
+        }
+
+        @Test
+        public void transportResponsePolicyWithRequirementCountReturnsSameValue() throws Exception {
+            AuthPolicy policy = new AuthPolicy(1, null, null, null, null);
+            when(transportResponse.getAuthPolicy()).thenReturn(policy);
+            AuthorizationResponse response = client.getAuthorizationResponse(authRequestId.toString());
+            assertEquals(Integer.valueOf(1), response.getPolicy().getRequiredFactors());
+        }
+
+        @Test
+        public void transportResponsePolicyWithRequirementCountReturnsNullForInherenceRequired() throws Exception {
+            AuthPolicy policy = new AuthPolicy(1, null, null, null, null);
+            when(transportResponse.getAuthPolicy()).thenReturn(policy);
+            AuthorizationResponse response = client.getAuthorizationResponse(authRequestId.toString());
+            assertNull(response.getPolicy().isInherenceFactorRequired());
+        }
+
+        @Test
+        public void transportResponsePolicyWithRequirementCountReturnsNullForKnowledgeRequired() throws Exception {
+            AuthPolicy policy = new AuthPolicy(1, null, null, null, null);
+            when(transportResponse.getAuthPolicy()).thenReturn(policy);
+            AuthorizationResponse response = client.getAuthorizationResponse(authRequestId.toString());
+            assertNull(response.getPolicy().isKnowledgeFactorRequired());
+        }
+
+        @Test
+        public void transportResponsePolicyWithRequirementCountReturnsNullForPossessionRequired() throws Exception {
+            AuthPolicy policy = new AuthPolicy(1, null, null, null, null);
+            when(transportResponse.getAuthPolicy()).thenReturn(policy);
+            AuthorizationResponse response = client.getAuthorizationResponse(authRequestId.toString());
+            assertNull(response.getPolicy().isPossessionFactorRequired());
+        }
+
+        @Test
+        public void transportResponsePolicyWithTypeRequiredHasNullForRequirementCount() throws Exception {
+            AuthPolicy policy = new AuthPolicy(null, true, true, true, null);
+            when(transportResponse.getAuthPolicy()).thenReturn(policy);
+            AuthorizationResponse response = client.getAuthorizationResponse(authRequestId.toString());
+            assertNull(response.getPolicy().getRequiredFactors());
+        }
+
+        @Test
+        public void transportResponsePolicyWithInherenceRequiredReturnsInherenceRequiredWithSameValue() throws Exception {
+            AuthPolicy policy = new AuthPolicy(null, true, false, false, null);
+            when(transportResponse.getAuthPolicy()).thenReturn(policy);
+            AuthorizationResponse response = client.getAuthorizationResponse(authRequestId.toString());
+            assertTrue(response.getPolicy().isInherenceFactorRequired());
+        }
+
+        @Test
+        public void transportResponsePolicyWithKnowledgeRequiredReturnsKnowledgeRequiredWithSameValue() throws Exception {
+            AuthPolicy policy = new AuthPolicy(null, false, true, false, null);
+            when(transportResponse.getAuthPolicy()).thenReturn(policy);
+            AuthorizationResponse response = client.getAuthorizationResponse(authRequestId.toString());
+            assertTrue(response.getPolicy().isKnowledgeFactorRequired());
+        }
+
+        @Test
+        public void transportResponsePolicyWithPossessionRequiredReturnsPossessionRequiredWithSAmeValue() throws Exception {
+            AuthPolicy policy = new AuthPolicy(null, false, false, true, null);
+            when(transportResponse.getAuthPolicy()).thenReturn(policy);
+            AuthorizationResponse response = client.getAuthorizationResponse(authRequestId.toString());
+            assertTrue(response.getPolicy().isPossessionFactorRequired());
+        }
+    }
+
+    @RunWith(Parameterized.class)
+    public static class AuthMethodsTypeTests {
+        private final String input;
+        private final Type expectedOutput;
+
+        @Parameterized.Parameters()
+        public static Iterable<Object[]> data() {
+            return Arrays.asList(new Object[][] {
+                    { "FACE", Type.FACE },
+                    { "CIRCLE_CODE", Type.CIRCLE_CODE },
+                    { "FINGERPRINT", Type.FINGERPRINT },
+                    { "GEOFENCING", Type.GEOFENCING },
+                    { "LOCATIONS", Type.LOCATIONS },
+                    { "PIN_CODE", Type.PIN_CODE },
+                    { "WEARABLES", Type.WEARABLES },
+                    { "UNKNOWN", Type.OTHER },
+            });
+        }
+
+        public AuthMethodsTypeTests(String input, Type expectedOutput) {
+            this.input = input;
+            this.expectedOutput = expectedOutput;
+        }
+
+        @Test
+        public void properlyMapsType() throws Exception {
+            UUID authRequestId = UUID.randomUUID();
+            UUID serviceId = UUID.randomUUID();
+            Transport transport = mock(Transport.class);
+            ServiceV3AuthsGetResponse transportResponse = mock(ServiceV3AuthsGetResponse.class);
+            when(transport.serviceV3AuthsGet(any(UUID.class), any(EntityIdentifier.class))).thenReturn(transportResponse);
+            when(transportResponse.getAuthorizationRequestId()).thenReturn(authRequestId);
+            when(transportResponse.getServicePins()).thenReturn(new String[]{});
+            when(transportResponse.getAuthMethods()).thenReturn(new AuthMethod[] {
+                    new AuthMethod(this.input, null, null, null, null, null, null, null)
+            });
+            BasicServiceClient client = new BasicServiceClient(serviceId, transport);
+            AuthorizationResponse response = client.getAuthorizationResponse(UUID.randomUUID().toString());
+            assertEquals(this.expectedOutput, response.getAuthMethods().get(0).getMethod());
+        }
+    }
+
+    @RunWith(MockitoJUnitRunner.class)
+    public static class AuthMethodsTests {
+
+        @Mock
+        private Transport transport;
+
+        @Mock
+        private ServiceV3AuthsGetResponse transportResponse;
+
+        private BasicServiceClient client;
+
+        private UUID serviceId = UUID.randomUUID();
+
+        private UUID authRequestId = UUID.randomUUID();
+
+        @Before
+        public void setUp() throws Exception {
+            when(transport.serviceV3AuthsGet(any(UUID.class), any(EntityIdentifier.class))).thenReturn(transportResponse);
+            when(transportResponse.getAuthorizationRequestId()).thenReturn(authRequestId);
+            when(transportResponse.getServicePins()).thenReturn(new String[]{});
+            client = new BasicServiceClient(serviceId, transport);
+        }
+
+        @Test
+        public void testNullReturnsNull() throws Exception {
+            AuthorizationResponse response = client.getAuthorizationResponse(UUID.randomUUID().toString());
+            assertNull(response.getAuthMethods());
+        }
+
+        @Test
+        public void testEmptyArrayReturnsEmptyList() throws Exception {
+            when(transportResponse.getAuthMethods()).thenReturn(new AuthMethod[]{});
+            AuthorizationResponse response = client.getAuthorizationResponse(UUID.randomUUID().toString());
+            assertNotNull(response.getAuthMethods());
+            assertThat(response.getAuthMethods(), is(empty()));
+        }
+    }
+
+    @RunWith(Parameterized.class)
+    public static class AuthMethodsAttributeValueTests {
+        private final Boolean value;
+
+        @Parameterized.Parameters()
+        public static Iterable<Object[]> data() {
+            return Arrays.asList(new Object[][] {
+                    { null },
+                    { Boolean.TRUE },
+                    { Boolean.TRUE },
+            });
+        }
+
+        private Transport transport;
+
+        private ServiceV3AuthsGetResponse transportResponse;
+
+        private BasicServiceClient client;
+
+        private UUID serviceId = UUID.randomUUID();
+
+        private UUID authRequestId = UUID.randomUUID();
+
+        public AuthMethodsAttributeValueTests(Boolean vaue) {
+            this.value = vaue;
+        }
+
+        @Before
+        public void setUp() throws Exception {
+            transportResponse = mock(ServiceV3AuthsGetResponse.class);
+            when(transportResponse.getAuthorizationRequestId()).thenReturn(authRequestId);
+            when(transportResponse.getServicePins()).thenReturn(new String[]{});
+            transport = mock(Transport.class);
+            when(transport.serviceV3AuthsGet(any(UUID.class), any(EntityIdentifier.class))).thenReturn(transportResponse);
+            client = new BasicServiceClient(serviceId, transport);
+        }
+
+        @Test
+        public void testIsSet() throws Exception {
+            when(transportResponse.getAuthMethods()).thenReturn(new AuthMethod[]{
+                    new AuthMethod(null, value, null, null, null, null, null, null)
+            });
+            AuthorizationResponse response = client.getAuthorizationResponse(authRequestId.toString());
+            assertEquals(value, response.getAuthMethods().get(0).getSet());
+        }
+
+        @Test
+        public void testIsActive() throws Exception {
+            when(transportResponse.getAuthMethods()).thenReturn(new AuthMethod[]{
+                    new AuthMethod(null, null, value, null, null, null, null, null)
+            });
+            AuthorizationResponse response = client.getAuthorizationResponse(authRequestId.toString());
+            assertEquals(value, response.getAuthMethods().get(0).getActive());
+        }
+
+        @Test
+        public void testIsAllowed() throws Exception {
+            when(transportResponse.getAuthMethods()).thenReturn(new AuthMethod[]{
+                    new AuthMethod(null, null, null, value, null, null, null, null)
+            });
+            AuthorizationResponse response = client.getAuthorizationResponse(authRequestId.toString());
+            assertEquals(value, response.getAuthMethods().get(0).getAllowed());
+        }
+
+        @Test
+        public void testIsSupported() throws Exception {
+            when(transportResponse.getAuthMethods()).thenReturn(new AuthMethod[]{
+                    new AuthMethod(null, null, null, null, value, null, null, null)
+            });
+            AuthorizationResponse response = client.getAuthorizationResponse(authRequestId.toString());
+            assertEquals(value, response.getAuthMethods().get(0).getSupported());
+        }
+
+        @Test
+        public void testIsUserRequired() throws Exception {
+            when(transportResponse.getAuthMethods()).thenReturn(new AuthMethod[]{
+                    new AuthMethod(null, null, null, null, null, value, null, null)
+            });
+            AuthorizationResponse response = client.getAuthorizationResponse(authRequestId.toString());
+            assertEquals(value, response.getAuthMethods().get(0).getUserRequired());
+        }
+
+        @Test
+        public void testIsPassed() throws Exception {
+            when(transportResponse.getAuthMethods()).thenReturn(new AuthMethod[]{
+                    new AuthMethod(null, null, null, null, null, null, value, null)
+            });
+            AuthorizationResponse response = client.getAuthorizationResponse(authRequestId.toString());
+            assertEquals(value, response.getAuthMethods().get(0).getPassed());
+        }
+
+        @Test
+        public void testIsError() throws Exception {
+            when(transportResponse.getAuthMethods()).thenReturn(new AuthMethod[]{
+                    new AuthMethod(null, null, null, null, null, null, null, value)
+            });
+            AuthorizationResponse response = client.getAuthorizationResponse(authRequestId.toString());
+            assertEquals(value, response.getAuthMethods().get(0).getError());
+        }
+    }
+
+
+    @RunWith(MockitoJUnitRunner.class)
+    public static class PolicyLocationsTests {
+
+        @Mock
+        private Transport transport;
+
+        @Mock
+        private ServiceV3AuthsGetResponse transportResponse;
+
+        @Mock
+        private AuthPolicy policy;
+
+        private BasicServiceClient client;
+
+        private UUID serviceId = UUID.randomUUID();
+
+        private UUID authRequestId = UUID.randomUUID();
+
+        @Before
+        public void setUp() throws Exception {
+            when(transport.serviceV3AuthsGet(any(UUID.class), any(EntityIdentifier.class))).thenReturn(transportResponse);
+            when(transportResponse.getAuthorizationRequestId()).thenReturn(authRequestId);
+            when(transportResponse.getServicePins()).thenReturn(new String[]{});
+            when(transportResponse.getAuthPolicy()).thenReturn(policy);
+            client = new BasicServiceClient(serviceId, transport);
+        }
+
+        @Test
+        public void testNullReturnsEmptyList() throws Exception {
+            AuthorizationResponse response = client.getAuthorizationResponse(UUID.randomUUID().toString());
+            assertThat(response.getPolicy().getLocations(), is(empty()));
+        }
+
+        @Test
+        public void testEmptyArrayReturnsEmptyList() throws Exception {
+            when(policy.getGeoFences()).thenReturn(new ArrayList<AuthPolicy.Location>());
+            AuthorizationResponse response = client.getAuthorizationResponse(UUID.randomUUID().toString());
+            assertThat(response.getPolicy().getLocations(), is(empty()));
+        }
+
+        @Test
+        public void testNullNameTransfers() throws Exception {
+            when(policy.getGeoFences()).thenReturn(Arrays.asList(new AuthPolicy.Location(null, 1.0, 2.0, 3.0)));
+            AuthorizationResponse response = client.getAuthorizationResponse(UUID.randomUUID().toString());
+            assertNull(response.getPolicy().getLocations().get(0).getName());
+        }
+
+        @Test
+        public void testNameTransfers() throws Exception {
+            when(policy.getGeoFences()).thenReturn(Arrays.asList(new AuthPolicy.Location("name", 1.0, 2.0, 3.0)));
+            AuthorizationResponse response = client.getAuthorizationResponse(UUID.randomUUID().toString());
+            assertEquals("name", response.getPolicy().getLocations().get(0).getName());
+        }
+
+        @Test
+        public void testRadiusTransfers() throws Exception {
+            when(policy.getGeoFences()).thenReturn(Arrays.asList(new AuthPolicy.Location("name", 1.0, 2.0, 3.0)));
+            AuthorizationResponse response = client.getAuthorizationResponse(UUID.randomUUID().toString());
+            assertEquals(1.0, response.getPolicy().getLocations().get(0).getRadius());
+        }
+
+        @Test
+        public void testLatitudeTransfers() throws Exception {
+            when(policy.getGeoFences()).thenReturn(Arrays.asList(new AuthPolicy.Location("name", 1.0, 2.0, 3.0)));
+            AuthorizationResponse response = client.getAuthorizationResponse(UUID.randomUUID().toString());
+            assertEquals(2.0, response.getPolicy().getLocations().get(0).getLatitude());
+        }
+
+        @Test
+        public void testLongitudeTransfers() throws Exception {
+            when(policy.getGeoFences()).thenReturn(Arrays.asList(new AuthPolicy.Location("name", 1.0, 2.0, 3.0)));
+            AuthorizationResponse response = client.getAuthorizationResponse(UUID.randomUUID().toString());
+            assertEquals(3.0, response.getPolicy().getLocations().get(0).getLongitude());
+        }
+    }
 }